XML and JSON output seperation

+++ Models/Opdracht.cs
@@ -3,7 +3,7 @@ public class Opdracht
     public string? RelatieCode { get; set; }
     public int? Aantal { get; set; }
     public string? ContainerType { get; set; }
-    public int? BrutoGewicht { get; set; }
+    public decimal? BrutoGewicht { get; set; }
 
     public List<string>? ZegelNummers { get; set; }
     public List<Activiteit>? Activiteiten { get; set; }
